requirements.txt
numpy~=1.21.5
pytorch
albumentations~=1.3.0
pillow~=9.2.0
tqdm~=4.64.1
torchvision~=0.14.0